# plugins/rest-api/routes/campaigns.routes.js
import { Router } from "express";
import multer from "multer";
import { apiAuthMiddleware, asyncHandler } from "../middleware/auth.middleware.js";
import { db } from "../../../server/db.js";
import { campaigns, contacts, agents, phoneNumbers, users } from "../../../shared/schema.js";
import { eq, and, desc, sql } from "drizzle-orm";
import { z } from "zod";
import { contactUploadService } from "../../../server/services/contact-upload-service.js";
const csvUpload = multer({
  storage: multer.memoryStorage(),
  limits: { fileSize: 10 * 1024 * 1024 },
  // 10MB limit
  fileFilter: (_req, file, cb) => {
    if (file.mimetype === "text/csv" || file.originalname.endsWith(".csv")) {
      cb(null, true);
    } else {
      cb(new Error("Only CSV files are allowed"));
    }
  }
});
const router = Router();
const createCampaignSchema = z.object({
  name: z.string().min(1, "Name is required").max(255),
  agentId: z.string().uuid("Invalid agent ID"),
  phoneNumberId: z.string().uuid().optional(),
  engine: z.enum(["elevenlabs", "plivo", "twilio-openai"]).optional(),
  scheduledStartTime: z.string().datetime().optional(),
  timezone: z.string().optional(),
  callWindowStart: z.string().regex(/^\d{2}:\d{2}$/).optional(),
  callWindowEnd: z.string().regex(/^\d{2}:\d{2}$/).optional(),
  maxConcurrentCalls: z.number().int().min(1).max(100).optional(),
  retryAttempts: z.number().int().min(0).max(5).optional(),
  retryDelayMinutes: z.number().int().min(1).max(1440).optional()
});
const addContactsSchema = z.object({
  contacts: z.array(z.object({
    phoneNumber: z.string().min(10),
    firstName: z.string().optional(),
    lastName: z.string().optional(),
    email: z.string().email().optional(),
    customFields: z.record(z.string()).optional()
  })).min(1).max(1e4)
});
router.post(
  "/",
  apiAuthMiddleware("campaigns:write"),
  csvUpload.single("contacts"),
  asyncHandler(async (req, res) => {
    const { userId } = req.apiAuth;
    let campaignData;
    if (req.body.data) {
      try {
        campaignData = typeof req.body.data === "string" ? JSON.parse(req.body.data) : req.body.data;
      } catch {
        const response2 = {
          success: false,
          error: { code: "VALIDATION_ERROR", message: "Invalid JSON in data field" },
          meta: { requestId: req.requestId, timestamp: (/* @__PURE__ */ new Date()).toISOString() }
        };
        return res.status(400).json(response2);
      }
    } else {
      campaignData = {
        name: req.body.name,
        agentId: req.body.agentId,
        phoneNumberId: req.body.phoneNumberId,
        engine: req.body.engine,
        scheduledStartTime: req.body.scheduledStartTime,
        timezone: req.body.timezone,
        callWindowStart: req.body.callWindowStart,
        callWindowEnd: req.body.callWindowEnd,
        maxConcurrentCalls: req.body.maxConcurrentCalls ? parseInt(req.body.maxConcurrentCalls) : void 0,
        retryAttempts: req.body.retryAttempts ? parseInt(req.body.retryAttempts) : void 0,
        retryDelayMinutes: req.body.retryDelayMinutes ? parseInt(req.body.retryDelayMinutes) : void 0
      };
    }
    const parseResult = createCampaignSchema.safeParse(campaignData);
    if (!parseResult.success) {
      const response2 = {
        success: false,
        error: {
          code: "VALIDATION_ERROR",
          message: "Invalid request body",
          details: { errors: parseResult.error.flatten().fieldErrors }
        },
        meta: { requestId: req.requestId, timestamp: (/* @__PURE__ */ new Date()).toISOString() }
      };
      return res.status(400).json(response2);
    }
    const data = parseResult.data;
    const [agent] = await db.select().from(agents).where(and(eq(agents.id, data.agentId), eq(agents.userId, userId))).limit(1);
    if (!agent) {
      const response2 = {
        success: false,
        error: { code: "NOT_FOUND", message: "Agent not found or does not belong to you." },
        meta: { requestId: req.requestId, timestamp: (/* @__PURE__ */ new Date()).toISOString() }
      };
      return res.status(404).json(response2);
    }
    let fromPhoneId = data.phoneNumberId;
    if (!fromPhoneId) {
      const [userPhone] = await db.select().from(phoneNumbers).where(and(eq(phoneNumbers.userId, userId), eq(phoneNumbers.status, "active"))).limit(1);
      if (userPhone) {
        fromPhoneId = userPhone.id;
      }
    }
    const [campaign] = await db.insert(campaigns).values({
      userId,
      agentId: data.agentId,
      phoneNumberId: fromPhoneId,
      name: data.name,
      status: "draft",
      scheduledStartTime: data.scheduledStartTime ? new Date(data.scheduledStartTime) : void 0,
      timezone: data.timezone || "UTC",
      callWindowStart: data.callWindowStart || "09:00",
      callWindowEnd: data.callWindowEnd || "18:00",
      maxConcurrentCalls: data.maxConcurrentCalls || 5,
      retryAttempts: data.retryAttempts || 2,
      retryDelayMinutes: data.retryDelayMinutes || 30
    }).returning();
    let contactStats = null;
    if (req.file) {
      try {
        const fileContent = await contactUploadService.readFileContent(req.file);
        const parsedContacts = contactUploadService.parseContactsFromCSV(fileContent, campaign.id);
        if (parsedContacts.length > 0) {
          const validContacts = parsedContacts.filter((c) => c.phone && c.phone.trim().length >= 10);
          const invalidCount = parsedContacts.length - validContacts.length;
          let added = 0;
          let skipped = 0;
          for (const contact of validContacts) {
            const [existing] = await db.select().from(contacts).where(and(eq(contacts.campaignId, campaign.id), eq(contacts.phone, contact.phone))).limit(1);
            if (existing) {
              skipped++;
              continue;
            }
            await db.insert(contacts).values({
              campaignId: campaign.id,
              phone: contact.phone,
              firstName: contact.firstName,
              lastName: contact.lastName,
              email: contact.email,
              customFields: contact.customFields,
              status: "pending"
            });
            added++;
          }
          await db.update(campaigns).set({ totalContacts: sql`COALESCE(${campaigns.totalContacts}, 0) + ${added}` }).where(eq(campaigns.id, campaign.id));
          contactStats = {
            fileName: req.file.originalname,
            totalRows: parsedContacts.length,
            contactsAdded: added,
            contactsSkipped: skipped,
            invalidRows: invalidCount
          };
        }
      } catch (error) {
        console.error("[Campaigns API] CSV processing error:", error.message);
        contactStats = {
          fileName: req.file.originalname,
          error: "Failed to parse CSV file"
        };
      }
    }
    const response = {
      success: true,
      data: {
        id: campaign.id,
        name: campaign.name,
        status: campaign.status,
        agentId: campaign.agentId,
        phoneNumberId: campaign.phoneNumberId,
        scheduledStartTime: campaign.scheduledStartTime,
        timezone: campaign.timezone,
        callWindowStart: campaign.callWindowStart,
        callWindowEnd: campaign.callWindowEnd,
        maxConcurrentCalls: campaign.maxConcurrentCalls,
        createdAt: campaign.createdAt,
        ...contactStats && { contacts: contactStats }
      },
      meta: { requestId: req.requestId, timestamp: (/* @__PURE__ */ new Date()).toISOString() }
    };
    res.status(201).json(response);
  })
);
router.get(
  "/",
  apiAuthMiddleware("campaigns:read"),
  asyncHandler(async (req, res) => {
    const { userId } = req.apiAuth;
    const page = parseInt(req.query.page) || 1;
    const pageSize = Math.min(parseInt(req.query.pageSize) || 20, 100);
    const offset = (page - 1) * pageSize;
    const [campaignList, countResult] = await Promise.all([
      db.select().from(campaigns).where(eq(campaigns.userId, userId)).orderBy(desc(campaigns.createdAt)).limit(pageSize).offset(offset),
      db.select({ count: sql`count(*)` }).from(campaigns).where(eq(campaigns.userId, userId))
    ]);
    const totalItems = Number(countResult[0]?.count || 0);
    const response = {
      success: true,
      data: campaignList.map((c) => ({
        id: c.id,
        name: c.name,
        status: c.status,
        agentId: c.agentId,
        totalContacts: c.totalContacts,
        calledContacts: c.calledContacts,
        scheduledStartTime: c.scheduledStartTime,
        startedAt: c.startedAt,
        completedAt: c.completedAt,
        createdAt: c.createdAt
      })),
      meta: {
        requestId: req.requestId,
        timestamp: (/* @__PURE__ */ new Date()).toISOString(),
        pagination: {
          page,
          pageSize,
          totalItems,
          totalPages: Math.ceil(totalItems / pageSize),
          hasNext: page < Math.ceil(totalItems / pageSize),
          hasPrev: page > 1
        }
      }
    };
    res.json(response);
  })
);
router.get(
  "/:id",
  apiAuthMiddleware("campaigns:read"),
  asyncHandler(async (req, res) => {
    const { userId } = req.apiAuth;
    const { id } = req.params;
    const [campaign] = await db.select().from(campaigns).where(and(eq(campaigns.id, id), eq(campaigns.userId, userId))).limit(1);
    if (!campaign) {
      const response2 = {
        success: false,
        error: { code: "NOT_FOUND", message: "Campaign not found." },
        meta: { requestId: req.requestId, timestamp: (/* @__PURE__ */ new Date()).toISOString() }
      };
      return res.status(404).json(response2);
    }
    const response = {
      success: true,
      data: campaign,
      meta: { requestId: req.requestId, timestamp: (/* @__PURE__ */ new Date()).toISOString() }
    };
    res.json(response);
  })
);
router.post(
  "/:id/contacts",
  apiAuthMiddleware("campaigns:write"),
  asyncHandler(async (req, res) => {
    const { userId } = req.apiAuth;
    const { id } = req.params;
    const [campaign] = await db.select().from(campaigns).where(and(eq(campaigns.id, id), eq(campaigns.userId, userId))).limit(1);
    if (!campaign) {
      const response2 = {
        success: false,
        error: { code: "NOT_FOUND", message: "Campaign not found." },
        meta: { requestId: req.requestId, timestamp: (/* @__PURE__ */ new Date()).toISOString() }
      };
      return res.status(404).json(response2);
    }
    if (campaign.status !== "draft" && campaign.status !== "paused") {
      const response2 = {
        success: false,
        error: { code: "CONFLICT", message: "Cannot add contacts to an active or completed campaign." },
        meta: { requestId: req.requestId, timestamp: (/* @__PURE__ */ new Date()).toISOString() }
      };
      return res.status(409).json(response2);
    }
    const parseResult = addContactsSchema.safeParse(req.body);
    if (!parseResult.success) {
      const response2 = {
        success: false,
        error: {
          code: "VALIDATION_ERROR",
          message: "Invalid request body",
          details: { errors: parseResult.error.flatten().fieldErrors }
        },
        meta: { requestId: req.requestId, timestamp: (/* @__PURE__ */ new Date()).toISOString() }
      };
      return res.status(400).json(response2);
    }
    const { contacts: newContacts } = parseResult.data;
    let added = 0;
    let skipped = 0;
    for (const contact of newContacts) {
      const [existing] = await db.select().from(contacts).where(and(eq(contacts.campaignId, id), eq(contacts.phone, contact.phoneNumber))).limit(1);
      if (existing) {
        skipped++;
        continue;
      }
      await db.insert(contacts).values({
        campaignId: id,
        phone: contact.phoneNumber,
        firstName: contact.firstName || "Unknown",
        lastName: contact.lastName,
        email: contact.email,
        customFields: contact.customFields,
        status: "pending"
      });
      added++;
    }
    await db.update(campaigns).set({ totalContacts: sql`${campaigns.totalContacts} + ${added}` }).where(eq(campaigns.id, id));
    const response = {
      success: true,
      data: {
        campaignId: id,
        contactsAdded: added,
        contactsSkipped: skipped
      },
      meta: { requestId: req.requestId, timestamp: (/* @__PURE__ */ new Date()).toISOString() }
    };
    res.status(201).json(response);
  })
);
router.post(
  "/:id/start",
  apiAuthMiddleware("campaigns:write"),
  asyncHandler(async (req, res) => {
    const { userId } = req.apiAuth;
    const { id } = req.params;
    const [campaign] = await db.select().from(campaigns).where(and(eq(campaigns.id, id), eq(campaigns.userId, userId))).limit(1);
    if (!campaign) {
      const response2 = {
        success: false,
        error: { code: "NOT_FOUND", message: "Campaign not found." },
        meta: { requestId: req.requestId, timestamp: (/* @__PURE__ */ new Date()).toISOString() }
      };
      return res.status(404).json(response2);
    }
    if (campaign.status === "running") {
      const response2 = {
        success: false,
        error: { code: "CONFLICT", message: "Campaign is already running." },
        meta: { requestId: req.requestId, timestamp: (/* @__PURE__ */ new Date()).toISOString() }
      };
      return res.status(409).json(response2);
    }
    const [user] = await db.select().from(users).where(eq(users.id, userId)).limit(1);
    if (!user || user.credits < campaign.totalContacts) {
      const response2 = {
        success: false,
        error: { code: "INSUFFICIENT_CREDITS", message: "Insufficient credits to run this campaign." },
        meta: { requestId: req.requestId, timestamp: (/* @__PURE__ */ new Date()).toISOString() }
      };
      return res.status(402).json(response2);
    }
    await db.update(campaigns).set({
      status: "running",
      startedAt: /* @__PURE__ */ new Date()
    }).where(eq(campaigns.id, id));
    const response = {
      success: true,
      data: {
        campaignId: id,
        status: "running",
        message: "Campaign started successfully."
      },
      meta: { requestId: req.requestId, timestamp: (/* @__PURE__ */ new Date()).toISOString() }
    };
    res.json(response);
  })
);
router.post(
  "/:id/pause",
  apiAuthMiddleware("campaigns:write"),
  asyncHandler(async (req, res) => {
    const { userId } = req.apiAuth;
    const { id } = req.params;
    const [campaign] = await db.select().from(campaigns).where(and(eq(campaigns.id, id), eq(campaigns.userId, userId))).limit(1);
    if (!campaign) {
      const response2 = {
        success: false,
        error: { code: "NOT_FOUND", message: "Campaign not found." },
        meta: { requestId: req.requestId, timestamp: (/* @__PURE__ */ new Date()).toISOString() }
      };
      return res.status(404).json(response2);
    }
    if (campaign.status !== "running") {
      const response2 = {
        success: false,
        error: { code: "CONFLICT", message: "Campaign is not running." },
        meta: { requestId: req.requestId, timestamp: (/* @__PURE__ */ new Date()).toISOString() }
      };
      return res.status(409).json(response2);
    }
    await db.update(campaigns).set({ status: "paused" }).where(eq(campaigns.id, id));
    const response = {
      success: true,
      data: {
        campaignId: id,
        status: "paused",
        message: "Campaign paused successfully."
      },
      meta: { requestId: req.requestId, timestamp: (/* @__PURE__ */ new Date()).toISOString() }
    };
    res.json(response);
  })
);
router.post(
  "/:id/contacts-upload",
  apiAuthMiddleware("campaigns:write"),
  csvUpload.single("file"),
  asyncHandler(async (req, res) => {
    const { userId } = req.apiAuth;
    const { id } = req.params;
    if (!req.file) {
      const response = {
        success: false,
        error: { code: "VALIDATION_ERROR", message: 'No CSV file provided. Upload a file with field name "file".' },
        meta: { requestId: req.requestId, timestamp: (/* @__PURE__ */ new Date()).toISOString() }
      };
      return res.status(400).json(response);
    }
    const [campaign] = await db.select().from(campaigns).where(and(eq(campaigns.id, id), eq(campaigns.userId, userId))).limit(1);
    if (!campaign) {
      const response = {
        success: false,
        error: { code: "NOT_FOUND", message: "Campaign not found." },
        meta: { requestId: req.requestId, timestamp: (/* @__PURE__ */ new Date()).toISOString() }
      };
      return res.status(404).json(response);
    }
    if (campaign.status !== "draft" && campaign.status !== "paused") {
      const response = {
        success: false,
        error: { code: "CONFLICT", message: "Cannot add contacts to an active or completed campaign." },
        meta: { requestId: req.requestId, timestamp: (/* @__PURE__ */ new Date()).toISOString() }
      };
      return res.status(409).json(response);
    }
    try {
      const fileContent = await contactUploadService.readFileContent(req.file);
      const parsedContacts = contactUploadService.parseContactsFromCSV(fileContent, id);
      if (parsedContacts.length === 0) {
        const response2 = {
          success: false,
          error: { code: "VALIDATION_ERROR", message: "CSV file is empty or has no valid rows." },
          meta: { requestId: req.requestId, timestamp: (/* @__PURE__ */ new Date()).toISOString() }
        };
        return res.status(400).json(response2);
      }
      const validContacts = parsedContacts.filter((c) => c.phone && c.phone.trim().length >= 10);
      const invalidCount = parsedContacts.length - validContacts.length;
      let added = 0;
      let skipped = 0;
      for (const contact of validContacts) {
        const [existing] = await db.select().from(contacts).where(and(eq(contacts.campaignId, id), eq(contacts.phone, contact.phone))).limit(1);
        if (existing) {
          skipped++;
          continue;
        }
        await db.insert(contacts).values({
          campaignId: id,
          phone: contact.phone,
          firstName: contact.firstName,
          lastName: contact.lastName,
          email: contact.email,
          customFields: contact.customFields,
          status: "pending"
        });
        added++;
      }
      await db.update(campaigns).set({
        totalContacts: sql`${campaigns.totalContacts} + ${added}`
      }).where(eq(campaigns.id, id));
      const response = {
        success: true,
        data: {
          campaignId: id,
          fileName: req.file.originalname,
          totalRows: parsedContacts.length,
          contactsAdded: added,
          contactsSkipped: skipped,
          invalidRows: invalidCount
        },
        meta: { requestId: req.requestId, timestamp: (/* @__PURE__ */ new Date()).toISOString() }
      };
      res.status(201).json(response);
    } catch (error) {
      const response = {
        success: false,
        error: {
          code: "VALIDATION_ERROR",
          message: "Failed to parse CSV file."
        },
        meta: { requestId: req.requestId, timestamp: (/* @__PURE__ */ new Date()).toISOString() }
      };
      return res.status(400).json(response);
    }
  })
);
var campaigns_routes_default = router;
export {
  campaigns_routes_default as default
};